Implement first file-delivery tests using jest
